Projet_django/dashboard/custompage/models.py
from django.db import models
from ckeditor.fields import RichTextField
from django.utils.text import slugify

class customPage(models.Model):
    title = models.CharField(max_length=100)
    slug = models.SlugField(unique=True, blank=True, null=True)
    content = RichTextField()
    
    def save(self, *args, **kwargs):
        self.slug = slugify(self.title)
        super().save(*args, **kwargs)
        
    def __str__(self):
        return self.title
    
