Projet_django/dashboard/requirements_local.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
django-ckeditor==6.7.0
django-js-asset==2.1.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
Pillow==10.1.0
PyJWT==2.8.0
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
twilio==8.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.6.0
yarl==1.9.2